migrate gpu test builder and implement a unifying trait
