+++ crates/vm/Cargo.toml
@@ -86,3 +86,8 @@ cuda = [
     "dep:openvm-cuda-backend",
     "dep:openvm-cuda-builder",
 ]
+touchemall = [
+    "cuda", 
+    "openvm-circuit-primitives/touchemall",
+    "openvm-cuda-backend/touchemall",
+]
