+++ crates/vm/build.rs
@@ -18,5 +18,19 @@ fn main() {
 
         builder.emit_link_directives();
         builder.build();
+
+        #[cfg(any(test, feature = "test-utils"))]
+        {
+            let builder = CudaBuilder::new()
+                .include_from_dep("DEP_CUDA_COMMON_INCLUDE")
+                .include("../circuits/primitives/cuda/include")
+                .include("../circuits/poseidon2-air/cuda/include")
+                .watch("cuda/src/testing")
+                .library_name("tracegen_gpu_testing")
+                .files_from_glob("cuda/src/testing/**/*.cu");
+
+            builder.emit_link_directives();
+            builder.build();
+        }
     }
 }
