+++ crates/vm/src/arch/testing/execution/cuda.rs
@@ -7,12 +7,13 @@ use openvm_circuit::{
     },
     utils::next_power_of_two_or_zero,
 };
-use openvm_stark_backend::{prover::types::AirProvingContext, Chip, ChipUsageGetter};
-use stark_backend_gpu::{
-    base::DeviceMatrix, cuda::copy::MemCopyH2D, prover_backend::GpuBackend, types::F,
+use openvm_cuda_backend::{
+    base::DeviceMatrix, chip::get_empty_air_proving_ctx, prover_backend::GpuBackend, types::F,
 };
+use openvm_cuda_common::copy::MemCopyH2D;
+use openvm_stark_backend::{prover::types::AirProvingContext, Chip, ChipUsageGetter};
 
-use crate::testing::cuda::execution_testing;
+use crate::cuda_abi::execution_testing;
 
 pub struct DeviceExecutionTester(pub(crate) ExecutionTester<F>);
 
@@ -54,11 +55,7 @@ impl<RA> Chip<RA, GpuBackend> for DeviceExecutionTester {
         let width = self.0.trace_width();
 
         if height == 0 {
-            return AirProvingContext {
-                cached_mains: vec![],
-                common_main: None,
-                public_values: vec![],
-            };
+            return get_empty_air_proving_ctx();
         }
         let trace = DeviceMatrix::<F>::with_capacity(height, width);
