+++ crates/vm/src/system/cuda/public_values.rs
@@ -120,6 +120,7 @@ mod tests {
     use rand::Rng;
 
     use super::PublicValuesChipGPU;
+    use crate::arch::testing::{GpuChipTestBuilder, GpuTestChipHarness, TestBuilder};
 
     type Harness = GpuTestChipHarness<
         F,
